requirements.txt
groq>=0.11.0
streamlit==1.35.0
pdfplumber==0.11.0
python-dotenv==1.0.1
